added Orders

+++ backend/Entities/Models/Repair.cs
@@ -5,7 +5,7 @@ namespace Entities.Models
     public class Repair : Order
     {
         [Required]
-        public int TotalPartCost { get; set; }
+        public decimal TotalPartCost { get; set; }
         public virtual IEnumerable<RepairPart> RepairParts { get; set; }
     }
 }
